Removed the filocommand spec , added the test to the filocli test
